refactor(cli): isolate gateway envelope parser from runtime imports
